requirements.txt
scrapy
SQLAlchemy
psycopg2
django
django-celery
redis
djangorestframework